Reword IntellisenseFacts and SnippetFacts to use TestOmniSharpHost
